menuService/requirements.txt
# ==========================================
# Production WSGI Server
# ==========================================
gunicorn==21.2.0

# ==========================================
# Django & API Framework
# ==========================================
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1

# ==========================================
# Authentication & Security
# ==========================================
djangorestframework-simplejwt==5.4.0

# ==========================================
# Database & Cache
# ==========================================
django-redis==5.4.0
psycopg2-binary==2.9.9
redis==5.0.1

# ==========================================
# Message Broker
# ==========================================
confluent-kafka==2.3.0

# ==========================================
# API Documentation (Swagger)
# ==========================================
drf-spectacular==0.27.2

# ==========================================
# Utilities & Configuration
# ==========================================
python-decouple==3.8

# ==========================================
# Testing
# ==========================================
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0